move previous research to r1
