+++ research/r1/README.md
