+++ research/r1/clickhouse/docker-compose.yml
