+++ research/r1/vertica/README.md
