+++ research/r1/vertica/config.py
