+++ research/r1/vertica/docker-compose.yml
