+++ research/r1/vertica/insert_test_results.txt
