+++ research/r1/vertica/select_test_results.txt
