+++ research/r1/vertica/select_tests.py
@@ -2,7 +2,7 @@ import random
 
 import vertica_python
 
-from config import CONNECTION_INFO
+from research.r1.vertica.config import CONNECTION_INFO
 from utils import calculate_avg, measure_exec_time, select_movie_ids, select_user_ids
 
 ITERATIONS = 200
