+++ research/r1/vertica/utils.py
